Add named slot support.
In Vue3, named slot are used with functions. JSX does not support this
case correctly. So VTYX handle children with a slot attribute and
instead of returning a VNode, it returns an object with a function which
returns the VNode.

This task open a work to support dynamic Scoped Slots
(https://v3.vuejs.org/guide/component-slots.html#scoped-slots).

It is possible that there are still some child that triggers a warning
that VNode are not functions: the ones which does not have slot
attribute. This should be fixed later (because it is currently only a
warning).

Change-Id: Id895c44f29963ec05af109ca69173e8e41e982f0