Clean child arguments.
In the h function, the third argument is related to child elements.
However sometime they have undefined values (or are set to false) while
a VNode is expected.

So before sending this list to Vue, the argument list is filtered.

Sometime an array is given while it should be flatten.

Change-Id: Iceee47e1fee3e4f19c6d6d1cca5e0ed6298f8db2