Add types for SVG elements.
It was not possible to use svg elements directly. So the jsx type
description has been extended for them.

It is still a great inspiration from @types/react (but from newer
version than it was before). But type description in React was very
non-precise and contains some mistakes.
So there are a little bit more precision for some elements. Some missing
elements have also been added (such as <set> which is not deprecated).

Deprecated attribute have been marked with @deprecated.

Deprecated elements are not supported (such as in the React file).

Change-Id: Iaa7b6b6ccd3c8c9a539b925aea699694031b4664